refactor: add validations package
