+++ packages/validations/src/func-result-type.js
