+++ packages/validations/src/index.js
@@ -19,3 +19,8 @@ export default function validateAST(ast: Program) {
     throw new Error(errorMessage);
   }
 }
+
+export { isConst } from "./is-const";
+export { getType, typeEq } from "./type-inference";
+
+export const stack = stackValidate;
