+++ packages/validations/src/is-const.js
