+++ packages/validations/test/fixtures/binop/module.wast
