+++ packages/validations/test/fixtures/br/module.wast
