+++ packages/validations/test/fixtures/load/module.wast
