+++ packages/validations/test/fixtures/store/module.wast
