+++ packages/webassemblyjs/test/index.js
@@ -5,43 +5,14 @@ const chai = require("chai");
 const { readFileSync } = require("fs");
 const path = require("path");
 const vm = require("vm");
-const { parse } = require("../../wast-parser/lib/");
-const assert = require("assert");
 
 const WebAssembly = require("../lib");
-const validate = require("../lib/compiler/validation/stack").default;
 
 function toArrayBuffer(buf) {
   return buf.buffer.slice(buf.byteOffset, buf.byteOffset + buf.byteLength);
 }
 
 describe("interpreter", () => {
-  describe("validation", () => {
-    const testSuites = glob.sync(
-      "packages/webassemblyjs/test/validation/**/module.wast"
-    );
-
-    testSuites.forEach(suite => {
-      describe(suite, () => {
-        const module = readFileSync(suite, "utf8");
-        const expectedErrors = readFileSync(
-          path.join(path.dirname(suite), "throws.txt"),
-          "utf8"
-        )
-          .split("\n")
-          .map(s => s.trim())
-          .filter(s => s.length > 0);
-
-        it("should give correct list of type errors", () => {
-          const ast = parse(module);
-          const errors = validate(ast);
-
-          assert.deepEqual(errors, expectedErrors);
-        });
-      });
-    });
-  });
-
   describe("wasm", () => {
     const testSuites = glob.sync(
       "packages/webassemblyjs/test/fixtures/**/module.wasm"
